pyproject.toml
[project]
name = "research-based-cancer-detection-with-explainable-ai"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "huggingface-hub>=1.11.0",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.2.0",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.26",
    "streamlit>=1.56.0",
    "tensorflow>=2.21.0",
    "uvicorn>=0.46.0",
]
